More tests for result type
