--- crates/cli/tests/reference/result-string.d.ts
@@ -1,3 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-export function exported(): string;
