+++ crates/cli/tests/reference/result.d.ts
@@ -1,0 +1,5 @@
+/* tslint:disable */
+/* eslint-disable */
+export function result_string(): string;
+export function result_void(): void;
+export function result_i32(): number;
